fix(rds): customizing secret results in unusable password and lost attachment
When the `excludeCharacters` prop is updated the secret is correctly
regenerated but the database is not updated because the
`MasterUserPassword` prop does not change. Moreover the connection
fields created by the attachment are lost because the secret is
regenerated but the `AWS::SecretsManager::SecretAttachment` doesn't
"rerun".

Introduce a new `fromFixedUsername()` static method in `Credentials`.
When used the `MasterUsername` prop of the database references the
username as a string instead of a dynamic reference to the username
field of the secret. Moreover the logical id of the secret is a hash of
its customization options. This way the secret gets replaced when it's
customized, the database gets updated and the attachement reruns. This
without updating the `MasterUsername` prop, avoiding a replacement of
the database.

Closes #11040