refactor(core): introduce cached lazies
`Lazy.stringValue()` used to be strictly uncached; however, this has an
extremely negative impact on synthesis performance when token resolution
is performed many times and complex operations are being performed in
Lazies, such as JSONification (see #11250 for context).

99% of the time, the value calculated by a Lazy does not actually
need to be recalculated, and the value can simply be cached. Right now,
we have no API to indicate whether or not caching on the return value is
okay.

Although it should *probably* be fine to switch the default behavior of
`Lazy.stringValue()` to be cached, it's still a little risky.

This PR introduces new methods to construct lazies:

* `Lazy.string()`/`Lazy.uncachedString()`.
* `Lazy.number()`/`Lazy.uncachedNumber()`.
* `Lazy.list()`/`Lazy.uncachedList()`.
* `Lazy.any()`/`Lazy.uncachedAny()`.

Which replace the existing `Lazy.xxxValue()`.

New method names were chosen instead of an additional option for
readability: the syntactic structure of how `Lazy.stringValue()` is
usually used makes adding a `{ cached: true }` options block to the end
of every existing call site rather visually noisy. Plus, deprecating the
existing method and replacing it with new ones forces current usage
sites to think clearly about whether or not their current use is
cacheable or not.

The *very* computationally heavy implemention of `toJsonString()` should
be changed to be cacheable, but has not been done yet in this PR. This
PR simply introduces the concept of cacheable vs uncacheable tokens.