fix(core): multiple library copies lead to 'Assets must be defined within Stage or App' error (#11113)
A recent change surfaces an issue where Stage objects from different
code copies don't recognize each other as a Stage. This paht aligns
the way how a Stage determines if something is a Stage to the same
mechanism that App and Stack use.

Thanks to @Shogan and @jogold for narrowing the issue down

fixes #10314


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*