fix(stepfunctions-tasks): incorrect S3 permissions for AthenaStartQueryExecution (#11203)
The changes made by https://github.com/aws/aws-cdk/pull/11045 grant S3 scoped permissions if the optional parameter  output location is passed. The output location is not parsed correctly to be attached as a resource. When the output location is correctly parsed, state machines with a valid definition and a valid S3 bucket still fail due to an `Unable to verify/create output bucket` error. The exact same state machine and S3 bucket pass with wildcard permissions as such the resource for Start Query Execution must be changed to `*`.

BREAKING CHANGE: type of `outputLocation` in the experimental Athena `StartQueryExecution` has been changed to `s3.Location` from `string`

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*