fix(stepfunctions): stack overflow when referenced json path finding encounters a circular object graph (#11225)
This change fixes #9319 where if an EcsRunTask step was used with the WAIT_FOR_TASK_TOKEN integration pattern, the logic for finding all referenced JSONPath elements in container overrides would encounter a circular reference and terminate with a stack overflow. 

With this change, `JsonPath.recurseObject` and `JsonPath.recurseArray` would maintain a list of already visited objects and use it to prune circular references back to an already visited object.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*