chore: fix update bump cfnspec script (#11302)
In #11191, we split up the running of the cfnspec update and the
adding files to git/commit step into sub-shells

However, variable assignments do not remain in effect after the subshell
completes. Currently, the `version` variable is not accessible when we
try to commit to Git.

Removed the execution of these steps in sub-shells as the separation is not
likely to provide a ton of benefit.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*