chore(core): update metadata tests to reflect v2 changes (#11320)
In v2, the `@aws-cdk/core` and `@aws-cdk/cx-api` modules are
private.
Update the tests so that it will only operate on v1. In v2, we will need
to design a new set of tests. This will be included in the re-design of
metadata collection - https://github.com/aws/aws-cdk-rfcs/pull/254


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*