fix(lambda): failed to add permission to an imported lambda from another account
Originally, when a function was imported into an account agnostic stack,
it was assumed that the function was from a different account and hence
its permission cannot be modified.

A subsequent change - 99111f72adc210f48e269db50f2b8e8b78d21252 - changed
this behaviour to its opposite. When an account agnostic stack was
encountered in the context of an imported function, it was assumed that
the function was part of the same account.
This has caused customers to report regressions - #11278, #11141.

This change reverts this behaviour back to its original assumption, with
the additional ability to configure this explicitly by the user if
needed for an advanced use case.

fixes #11141