chore(lambda-nodejs): drop coverage requirement
The `lambda-nodejs` package sometimes exhibits a coverage drop
on the build server, failing the build with the following error:

```
@aws-cdk/aws-lambda-nodejs: Jest: "global" coverage threshold for branches (60%) not met: 53.75%
```

Repeating the build again typically fixes the error.

I could not reproduce the issue locally, "on my machine" the build
just works with sufficient coverage.

The current behavior is a drain on PR builds, so I'm lowering the
coverage requirement so builds don't fail on unrelated changes.