chore: Create empty CHANGELOG.v2.md as placeholder
The `v2-main` branch currently doesn't build due to failures at the pack stage,
where it attempts to copy the CHANGELOG to the `dist` directory. For V2, the
CHANGELOG file is specified as `CHANGELOG.v2.md`, but that file doesn't (yet)
exist.

This creates the empty header section of the CHANGELOG so the build can pass.