chore: 2.0.0a0 is accidentally identified as 0.0.0 by pack script
The pack script defensively checks if there are any artifacts versioned 0.0.0 under the `dist/` directory. Python artifacts are named like so: `aws_cdk.aws_iotanalytics-2.0.0a0-py3-none-any.whl` and since by default grep uses regular expressions, this matches `grep 0.0.0`.

The fix is to use `-F` so that the search string is not treated as regex.