feat(core): natively support .dockerignore (#10922)
Patterns in `.dockerignore` used to be interpreted as globs, which has different behavior from how `.dockerignore` is supposed to be interpreted. Specifically it was hard to properly ignore whole directories at once (such as a large `node_modules`).

This PR adds an `ignoreMode` flag which controls how the ignore patterns should be interpreted. The choices are:

* `IgnoreMode.GLOB` - interpret as before
* `IgnoreMode.GIT` - interpret as in `.gitignore`
* `IgnoreMode.DOCKER` - interpret as in `.dockerignore`

Users might have dependencies on the current behavior, which is why the default `ignoreMode` depends on a feature flag: `@aws-cdk/aws-ecr-assets:dockerIgnoreSupport` (which is set to `true` for new projects) enables the new, correct Docker-ignore behavior by default.

----

Closes #10921

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*