fix(autoscaling): `targetRequestsPerSecond` is actually requests per minute (#11457)
Scaling to a request rate per instance was advertised to scale to a
rate/second, but the underlying API was actually expecting a
rate/minute.

Upscale the value of `targetRequestsPerSecond` by a factor of 60 to
make its name actually correct, deprecate it, and add a
`targetRequestsPerMinute` property which should have been the original
API.

Fixes #11446.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*