fix(pipelines): synthesizes incorrect paths on Windows (#11464)
Pipelines generates path references based on the local system's
file tree, and uses `path.join()` (etc) to build those.

On a Windows machine, those would use a `\\` as a path separator,
but the path separator for CodePipeline and the Linux CodeBuild image
we use should be a `/`.

Translate them.

Fixes #11359, fixes #11405, fixes #11424.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*