docs(apigateway): Augment API docs for LambdaIntegrationOptions proxy flag (#11400)
Setting the proxy flag for LambdaIntegrationOptions circumvents the need for an integration response. However a common mistake is to omit HTTP Headers from backend lambda services when enabling Cors or returning JSON. 

This PR includes a link to the AWS Developer guide which highlights it is the responsibility of the backend service to return these HTTP headers. 


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*