feat(events-targets): add CloudWatch LogGroup Target (#10598)
Implementation

Update package `@aws-cdk/aws-events-targets` to include support for `CloudWatch LogGroups`.
The `CloudWatchLogGroup` event target must add a resource policy to CloudWatch LogGroups to allow events to be posted to the LogGroup. It requires a `CustomResource` to do so as it's not supported by CloudFormation.

The `log-group-resource-policy.ts` file should be moved to another module related to LogGroups so it can be easily shared. At the time of this pull request, it is not possible to add it into the `aws-logs` module because of a circular dependency issue.


Closes #9953

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*