chore(cfnspec): race condition in test deletes specification.json (#11557)
The build script of `@aws-cdk/cfnspec` creates the file
`spec/specification.json` by running `node build-tools/build`. This file
runs an `async` `main()`.

The test file `test.build.ts` imports this same file to test one of its
utils (`massageSpec`). This leads to the script potentially running one
more time if the test is slow. And sometimes the test stops just when a
`spec/specification.json` is about to be written. The result is a zero
byte JSON file. In this case, the next package to be tested during the
build (`@aws-cdk/cloudformation-diff`) fails when it uses
`@aws-cdk/cfnspec` because it now requires a zero byte
`spec/specification.json`.

Move all the utils out of the build to allow importing them without
rerunning the specifications build.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*