chore(cloudfront): mark Distribution as Dev Preview in README
The main header and `package.json` were updated a while back, but this extra
section in the README was missed. This was discovered and fixed in #10500, but
that PR hasn't been merged yet (and may still not be for a while). Separating it
out into this PR to get it out faster.