fix(cli): credential provider plugins cannot be used with modern synthesis (#11350)
I'm not currently able to use CDK pipelines because I use a credential provider. I believe this is because the new bootstrap uses assumed roles. When assuming a role, only defaultCredentials is checked, I'm not sure if this was intentional or not because there is a comment that seems to mention this (that I will edit if this PR has any merit), but it seems odd to not allow a credential provider here.

I'm not 100% sure why I had to change the account ID in the tests, but I did and all the tests seem to still pass.