fix(cli): fix init-templates for v2
This change updates all of the init templates to be v2 compatible. This mostly
involves adjusting depenendencies and import statements.

One call-out -- we don't (yet) have a plan for `assert` for v2. This is being
tracked in the v2 project board, but is currently out of scope for the alpha
release. This means I had to rewrite several tests that used `assert` to doing
much simpler/uglier validations. I used the Java tests as the basis for these
rewrites.

Tested against a recent build artifact to verify correctness.