fix(eks): cluster creation fails when configured with an imported public subnet and private endpoint (#11620)
The problem was that we were checking for existence using the entire subnet object, instead of the subnet id.
This breaks when the subnet selection is configured with concrete `Subnet` objects, since the existence check was using object comparison, instead of string (id). 

Fixes https://github.com/aws/aws-cdk/issues/11471

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*