fix(core): reusing StackSynthesizer leads to unsynthesized Stacks (#11635)
If the same StackSynthesizer object is used for multiple stacks,
only the last one is properly synthesized to the cloud assembly.

Ideally, we would have made it so that there is a difference between
the specifying `StackSynthesizer` object and a "bound" version of
the `StackSynthesizer` (bound to a `Stack`).

Unfortunately, we cannot introduce that change without heavy
contortions in order not to break backwards compatibility, so the
simple thing to do right now is to make it blindingly obvious that
"you cannot do that".

Fixes #11528.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*