chore(cdk): init templates are major-version aware (#11665)
In preparation of non-trivial differences in the init templates between v1 and
v2 (see #11638), moving all of the existing init templates to a /v1 subdirectory
and making the init library major-version aware.

Note: This change only concerns the v1 template changes. The v2 init templates are
already broken, and I will adapt #11638 once this has been merged (and forward-
merged).

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*