fix(scripts): lerna not found in merge-forward
The pkglint command is failing on the merge-forward job due to lerna not being
found. This also wasn't failing the build itself, just logging an error.
Fixed the script to (a) run pkglint via yarn and (b) run as its own command so
the failed exit code causes the whole script to fail.