feat(cli): add `--no-lookups` flag to disable context lookups (#11489)
Context lookups are supposed to be performed on developer desktops, and
committed to `cdk.context.json`. If you don't, your CI build might try
to perform a lookup and fail with an unclear error message about
permissions, or worse: appear to work properly but leave you with a
nondeterministic build.

Introduce a CLI flag called `--no-lookups` that throws an appropriately
descriptive error message if you forgot to perform context lookups
before committing.

This now also makes it possible to write an integration test for
PR #11461.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*