feat(cloudfront): Lambda@Edge construct (#10500)
This PR creates a construct (`EdgeFunction`) for Lambda@Edge functions.
CloudFront requires that a function be in us-east-1 to be used with Lambda@Edge,
even if the logical distribution is created via another region. The initial goal
of this construct is to make it easier to request and work with a function in
us-east-1 when the primary stack is in another region.

In the future, this can be extended to validate other Lambda@Edge restrictions.
See https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/lambda-requirements-limits.html
and #9833 for more information on those limitations.

Thanks to @asterikx for the inspiration and consolidated writeup in #1575.

Related changes:
* When updating the CloudFront README, I noticed that the `Distribution` sub-section hadn't been updated when it was flipped from Experimental to Dev Preview.
* A minor docstring fix for Lambda.

fixes #9862


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*