chore(pkglint): rule to ensure heading consistency in README.md
Consistent heading structure in the `README.md` files is necessary in
order to ensure consistent rendering in the Python documentation, as the
rendering of the TOC tree is dependent on heading levels. The best
experience is achieved by having all `README.md` files start with a `H1`
heading, and subsequently contain only `H2` and deeper headings. It is
also important that no heading level is skipped (i.e: a `H4` "nested" in
a `H2`).

This adds a `pkglint` rule that parses the `README.md` files using the
`commonmark` parser, and ensures the heading structure is consistent per
the above rules, and registers fixable offenses otherwise.

Each offense is reported (and fixed) independently in an effort to
provide "more actionnable" feedback in the context of PR validation, and
to avoid risking to modify other parts of the document if we were to
simply "re-render" the updated Markdown.