chore: add "canned metrics" from CloudWatch Explorer's model
CloudWatch Explorer is a new service who have been collecting a
set of metrics emitted by various services.

Integrate this model and use to generate convenience methods that are
aware of per-metric dimensions and suggested default statistics.

We can switch over to these metrics to a get second source of
confirmation that we have the dimensions correct.