chore: init templates for v2 (#11706)
Creates the v2 init templates by first copying the v1 init templates wholesale,
and then applying a series of changes to make the templates v2-compatible.
This mostly involves adjusting dependencies and import statements.

For ease of review, I suggest looking at the second commit, which is just the
changes to the v2 templates from the v1 baseline.

One call-out -- we don't (yet) have a plan for assert for v2. This is being
tracked in the v2 project board, but is currently out of scope for the alpha
release. This means I had to rewrite several tests that used assert to doing
much simpler/uglier validations. I used the Java tests as the basis for these
rewrites.

Tested against a recent build artifact to verify correctness.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*