revert function hash changes
The original CFN patch updated `Runtime` and `Handler` fields (changed optionality). Because of how json-patch works and the fact that currently our lambda function hash calculation is order-sensitive,
it caused invalidation of all lambda hashes.

To avoid that, we use a marker value to signal that the value should be removed by an escape hatch. This should be reverted once the new CFN spec is published.