fix(lambda-nodejs): automatic entry finding with compilerOptions.outDir
When the defining file is a `.ts` that is emitted into an
[`outDir`](https://www.typescriptlang.org/tsconfig#outDir) extracting
the defining filename with the stacktrace returns the original path
because of the sourcemap.

So, if the `.ts` file is defined in `src` and emitted into `lib`
using the stacktrace returns the path in `src`. When shipping a
construct library with npm the `src` is not published and the
automatic entry finding fails.

Use the `CallSite` API to avoid this problem.

Inspired by https://github.com/sindresorhus/callsites and
https://github.com/sindresorhus/caller-callsite.