fix(lambda): changing reserved concurrency fails lambda version deployment
Function hash calculations include properties such as
`ReservedConcurrentExecutions`, `Tags` and `DependsOn` which do not
affect lambda versions. This change makes sure that only `Properties`
section of `AWS::Lambda::Function` is included in the hash calculation
and that specific properties are excluded from them.

Further, the hash calculation is sensitive to key order. This causes
the asset hash to change when any foundational change is made to the
CloudFormation spec or synthesizer, which then invalidates all customer
lambda functions. This can be avoided by making the hash resilient to
key order.

fixes #11537
closes #11707