chore: CHANGELOG is missing a breaking change entry for efs (#11745)
The change made in #11524 which switched from using the `keyId` to `keyArn` 
is a BREAKING change as an update requires replacement:
https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-efs-filesystem.html#cfn-efs-filesystem-kmskeyid

The `efs` module is experimental and this was a capability that needed to be fixed ahead
of moving to developer preview. It was missed that an update to the keyId replaces the 
filesystem. Amending the CHANGELOG to indicate that the filesystem will be replaced.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*