fix(pipelines): allow cdk stages to deploy more than 25 stacks
CdkStage can now create multiple codepipeline stages. If the stage
has not enough room remaining for the next stack, it will insert a new
stage and continue processing it. Manually added actions are also
added to the correct stage depending on their run order.

Add afterPreparation method to CdkStage to allow a function to be
executed after the CdkStage has been prepared (actions bound to their
underlying codepipeline stage).

Closes #11315

BREAKING CHANGE: IAction actions are bound to the underlying
codepipeline stages of the CdkStage late.
* **pipelines:** ShellScriptAction IGrantable interface can only be
used after the action is effectively bound to the stage, usage of
afterPreparation is required to use the action as an IGrantable