fix(lambda-nodejs): automatic entry finding with compilerOptions.outDir (#11729)
When the defining file is a `.ts` that is emitted into an
[`outDir`](https://www.typescriptlang.org/tsconfig#outDir) extracting
the defining filename with the stacktrace returns the original path
because of the sourcemap.

So, if the `.ts` file is defined in `src` and emitted into `lib`
using the stacktrace returns the path in `src`. When shipping a
construct library with npm the `src` is not published and the
automatic entry finding fails.

The `src`/`lib` architecture is the one in https://github.com/projen/projen.

Use the `CallSite` API to avoid this problem.

Inspired by https://github.com/sindresorhus/callsites and
https://github.com/sindresorhus/caller-callsite.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*