fix(core): init templates not initialized when running the CLI from source (#11731)
As part of the CDKv2 work, #11665 made the init templates version-aware
(allowing for different init templates for v1 and v2). This works when the CLI
\is run from a packaged distribution, but when run locally in development --
for example, when running `cdk-integ` -- the CLI was failing due to the local
version being 0.0.0, and the templates looking for a "v0" directory.

Fix the local experience by (currently) defaulting to v1 templates for local
development.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*