fix(ec2): instance not replacedwhen changing asset in UserData
The Logical ID hash used in Instance was based on the UserData string,
but it needed to be based on the *resolved* version of the UserData
string (since only the resolved version shows the asset hash).

Slight complication because the resolution of the UserData may refer
back to the LogicalID, which we need to account for by using a
recursion breaker.

Fixes #11704.