fix(core): cannot parse ARNs with `/` in resource name
Some resource ARNs have "path"-like names, with `/` in them.

We used to do simple splits on separators, which made it impossible
to properly parse entire paths. Example:

```js
'arn:aws:iam::111111111111:role/path/to/role/name'

// split on ':', take 6th element
'role/path/to/role/name'

// split on '/', take 1st or 2nd element
['role', 'path']
```

You see we can get `'role'` and we can get `'path'`, but we cannot get
`path/to/role/name` in this way.

However, once we have `role`, something else we can do is split on
`:role/` and take the 2nd element of the resulting array. This will be
the entire path after all!

```js
'arn:aws:iam::111111111111:role/path/to/role/name'

// split on `:role/`
['arn:aws:iam::111111111111', 'path/to/role/name']
```

We rely on the fact that the `:` anchors the split string to the
start of the path component; it's unlikely that a `:` is going to
occur anywhere else in an ARN.

The resulting expression is fiercely complex. Don't stare at it too
much or it might fry your brain. It does calculate the right value,
which is the most important thing:

```js
{
  'Fn::Select': [1, {
    'Fn::Split': [
      {
        'Fn::Join': ['', [
          ':',
          { 'Fn::Select': [0, { 'Fn::Split': ['/', { 'Fn::Select': [5, { 'Fn::Split': [':', theToken] }] }] }] },
          '/',
        ]],
      },
      theToken,
    ],
  }],
}
```

Fixes #11705