fix(iam): OIDC provider cannot be imported from parameter
The "resource name" part of the OIDC Provider ARN contains the
ARN separator (`/`), which cannot be properly split into components
using the primitives that CloudFormation gives us.

Introduce a new helper method on `Arn` which can obtain the
resource name, slashes and all, provided that we know the resource
type for a fact.

Fixes #11705.