fix(sqs): queueUrl property has incorrect region and account for imported queue (#11651)
fix(sqs): imported queue using incorrect options to generate queueUrl

Currently if we use `fromQueueArn` static helper to import an existing queue, the generated `queueUrl` property uses incorrect account and region. It uses the stack's region/account in which it is imported rather than the one in which it was created.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*