feat(cfn-include): preserve properties of resources that are not in the current CFN schema
Currently, if cloudformation-include encounters a property in one of the resources the CDK has an L1 for
that is not in the version of the CloudFormation resource schema the module uses,
the property will be silently dropped. Of course, that is not ideal.

Make changes to our L1 code generation to preserve all properties that are not in the CFN schema.

Fixes #9717