chore: fix integ tests leaking bootstrap buckets
The bootstrap default template was changed in #11547 to set the deletion policy
of the staging bucket to `Retain`. This has the impact of the integration tests
now accumulating buckets on each run, quickly exhausting the buckets/account
limits.

Explicitly delete the staging bucket after the stack is deleted so we stop
leaking buckets on each run.