chore(cdk-lint): add linting to README.md files (#11708)
Consistent heading structure in the `README.md` files is necessary in
order to ensure consistent rendering in the Python documentation, as the
rendering of the TOC tree is dependent on heading levels. The best
experience is achieved by having all `README.md` files start with a `H1`
heading, and subsequently contain only `H2` and deeper headings. It is
also important that no heading level is skipped (i.e: a `H4` "nested" in
a `H2`).

This adds a `cdk-lint` stage that executes `markdownlint` on the
`README.md` files, with a configuration that is crafted specifically to
maximize compatibility with most  Markdown parsers, and to ensure
correct rendering in the Python docs. Many of these offenses can be
automatically fixed by running `cdk-lint --fix` or `yarn lint --fix`, however
some cannot be automatically fixed by the tool and will require human
intervention.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*